+++ test/test_util.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from test.data import TEST_DATA_DIR
 from test.utils.graph import cached_graph
 from test.utils.namespace import RDFT
-from typing import Any, Collection, List, Optional, Set, Tuple, Type, Union
+from typing import Any, Callable, Collection, List, Optional, Set, Tuple, Type, Union
 
 import pytest
 
@@ -15,7 +15,7 @@ from rdflib import XSD, util
 from rdflib.graph import ConjunctiveGraph, Graph, QuotedGraph
 from rdflib.namespace import RDF, RDFS
 from rdflib.term import BNode, IdentifiedNode, Literal, Node, URIRef
-from rdflib.util import _coalesce, _iri2uri, find_roots, get_tree
+from rdflib.util import _coalesce, _convert_optional, _iri2uri, find_roots, get_tree
 
 n3source = """\
 @prefix : <http://www.w3.org/2000/10/swap/Primer#>.
@@ -347,19 +347,76 @@ class TestUtilTermConvert:
 
 
 @pytest.mark.parametrize(
-    ["params", "expected_result"],
+    ["params", "default", "expected_result"],
     [
-        ([], None),
-        (["something"], "something"),
-        ([False, "something"], False),
-        (["", "something"], ""),
-        ([0, "something"], 0),
-        ([None, "something", 1], "something"),
-        (["something", None, 1], "something"),
+        ([], ..., None),
+        (["something"], ..., "something"),
+        ([False, "something"], ..., False),
+        (["", "something"], ..., ""),
+        ([0, "something"], ..., 0),
+        ([None, "something", 1], ..., "something"),
+        (["something", None, 1], ..., "something"),
+        (["something", None, 1], 5, "something"),
+        ([], 5, 5),
+        ([None], 5, 5),
+        ([None, None], 5, 5),
+        ([None, None], 5, 5),
     ],
 )
-def test__coalesce(params: Collection[Any], expected_result: Any) -> None:
-    assert expected_result == _coalesce(*params)
+def test__coalesce(params: Collection[Any], default: Any, expected_result: Any) -> None:
+    if default == Ellipsis:
+        result = _coalesce(*params)
+    else:
+        result = _coalesce(*params, default)
+    assert expected_result == result
+
+
+def test__coalesce_typing() -> None:
+    """
+    type checking for _coalesce behaves as expected.
+    """
+    str_value: str
+    optional_str_value: Optional[str]
+
+    optional_str_value = _coalesce(None, "a", None)
+    assert optional_str_value == "a"
+
+    str_value = _coalesce(None, "a", None)  # type: ignore[assignment]
+    assert str_value == "a"
+
+    str_value = _coalesce(None, "a", None, default="3")
+    assert str_value == "a"
+
+
+@pytest.mark.parametrize(
+    ["converter", "input", "expected_result"],
+    [
+        (str, None, None),
+        (str, "5", "5"),
+        (str, 5, "5"),
+    ],
+)
+def test__convert_optional(
+    converter: Callable[[Any], Any], input: Any, expected_result: Any
+) -> None:
+    result = _convert_optional(converter, input)
+    assert expected_result == result
+
+
+def test__convert_optional_typing() -> None:
+    """
+    type checking for _convert_optional behaves as expected.
+    """
+    optional_str_value: Optional[str]
+
+    optional_str_value = _convert_optional(str, 1)
+    assert optional_str_value == "1"
+
+    optional_str_value = _convert_optional(str, "1")
+    assert optional_str_value == "1"
+
+    optional_str_value = _convert_optional(str, None)
+    assert optional_str_value is None
 
 
 @pytest.mark.parametrize(